src/requirements.txt
PyMuPDF
Pillow
GitPython
customtkinter
pywin32
